# backend/swagger/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Team CalTalk API",
    "description": "팀 단위 캘린더 기반 일정관리 + 채팅 애플리케이션 API. API Route 소스 위치: backend/app/api/",
    "version": "1.3.0",
    "x-source-directory": "backend/app/api"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Base URL (Next.js API Routes — backend/app/api/)"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string",
            "description": "에러 메시지"
          }
        }
      },
...
      "JoinRequest": {
        "type": "object",
        "required": ["id", "teamId", "teamName", "requesterId", "status", "requestedAt"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "e5f6a7b8-c9d0-1234-efab-cd5678901234"
          },
          "teamId": {
            "type": "string",
            "format": "uuid",
            "example": "b1c2d3e4-f5a6-7890-bcde-fa1234567890"
          },
          "teamName": {
            "type": "string",
            "example": "개발팀"
          },
          "requesterId": {
            "type": "string",
            "format": "uuid",
            "example": "d4e5f6a7-b8c9-0123-defa-bc3456789012"
          },
          "status": {
            "type": "string",
            "enum": ["PENDING", "APPROVED", "REJECTED"],
            "example": "PENDING"
          },
          "requestedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-04-08T09:00:00.000Z"
          },
          "respondedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "example": null
          }
        }
      },
      "JoinRequestDetail": {
        "allOf": [
          { "$ref": "#/components/schemas/JoinRequest" },
          {
            "type": "object",
            "required": ["requesterName"],
            "properties": {
              "requesterName": {
                "type": "string",
                "example": "김철수"
              },
              "requesterEmail": {
                "type": "string",
                "format": "email",
                "example": "kimcs@example.com"
              }
            }
          }
        ]
      },
...
    }
  },
  "paths": {
...
    "/teams/{teamId}/join-requests": {
      "post": {
        "tags": ["JoinRequests"],
        "summary": "팀 가입 신청 제출",
        "description": "로그인한 사용자가 특정 팀에 가입 신청을 제출합니다. TeamJoinRequest 레코드를 PENDING 상태로 생성하며, 해당 팀 팀장의 나의 할 일 목록에 자동으로 표시됩니다.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "가입 신청할 팀의 UUID"
          }
        ],
        "responses": {
          "201": {
            "description": "가입 신청 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinRequest"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "인증이 필요합니다." }
              }
            }
          },
          "404": {
            "description": "팀 없음",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "팀을 찾을 수 없습니다." }
              }
            }
          },
          "409": {
            "description": "충돌",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "already_member": {
                    "value": { "error": "이미 해당 팀의 구성원입니다." }
                  },
                  "pending_exists": {
                    "value": { "error": "이미 가입 신청이 진행 중입니다." }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["JoinRequests"],
        "summary": "팀 PENDING 가입 신청 목록 조회",
        "description": "특정 팀의 PENDING 상태 가입 신청 목록을 조회합니다. 팀장(LEADER)만 접근할 수 있습니다.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "조회할 팀의 UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "가입 신청 목록 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["teamId", "teamName", "joinRequests"],
                  "properties": {
                    "teamId": {
                      "type": "string",
                      "format": "uuid",
                      "example": "b1c2d3e4-f5a6-7890-bcde-fa1234567890"
                    },
                    "teamName": {
                      "type": "string",
                      "example": "개발팀"
                    },
                    "joinRequests": {
                      "type": "array",
                      "description": "PENDING 상태 가입 신청 배열 (requestedAt 오름차순 정렬)",
                      "items": {
                        "$ref": "#/components/schemas/JoinRequestDetail"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "인증이 필요합니다." }
              }
            }
          },
          "403": {
            "description": "권한 없음 (LEADER만 가능)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "팀장만 가입 신청 목록을 조회할 수 있습니다." }
              }
            }
          },
          "404": {
            "description": "팀 없음",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "팀을 찾을 수 없습니다." }
              }
            }
          }
        }
      }
    },
    "/teams/{teamId}/join-requests/{requestId}": {
      "patch": {
        "tags": ["JoinRequests"],
        "summary": "가입 신청 승인/거절",
        "description": "팀장이 PENDING 상태의 가입 신청을 승인(APPROVE) 또는 거절(REJECT)합니다. 승인 시 team_members에 MEMBER로 원자적 등록되고 status가 APPROVED로 갱신됩니다.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "소속 팀 UUID"
          },
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "처리할 가입 신청 UUID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["action"],
                "properties": {
                  "action": {
                    "type": "string",
                    "enum": ["APPROVE", "REJECT"],
                    "example": "APPROVE"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "가입 신청 처리 성공 (승인 또는 거절)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinRequestDetail"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "invalid_action": {
                    "value": { "error": "action은 APPROVE 또는 REJECT이어야 합니다." }
                  },
                  "missing_action": {
                    "value": { "error": "action은 필수입니다." }
                  },
                  "already_processed": {
                    "value": { "error": "이미 처리된 가입 신청입니다." }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "인증이 필요합니다." }
              }
            }
          },
          "403": {
            "description": "권한 없음 (LEADER만 가능)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "팀장만 가입 신청을 처리할 수 있습니다." }
              }
            }
          },
          "404": {
            "description": "리소스 없음",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "team_not_found": {
                    "value": { "error": "팀을 찾을 수 없습니다." }
                  },
                  "request_not_found": {
                    "value": { "error": "가입 신청을 찾을 수 없습니다." }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}